refactoring renamed GameManager into World
